Backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9
google-genai>=0.3
elevenlabs>=0.30
python-dotenv>=1.0
